General fixes and reorganization
Closes #4
